finish home page v:1.0.1 refactor JSX code
