+++ client/src/Pages/MainPages/SecondBlock/HotTour.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export const HotTour = () => {
+  return (
+    <div className='hotel'>
+      <div className='hotel-wrapper'>
+        <span className='city-name'>London</span>
+        <span className='count-hotels'>3200 Hotels</span>
+      </div>
+    </div>
+  )
+}
