+++ client/src/Pages/MainPages/ThirdBlock/ThirdBlock.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+export const ThirdBlock = () => {
+  return (
+    <section className='third'>
+      <div className='third-wrapper'>
+        <span className='third-title'>Safe Time And Money</span>
+        <p>Sign up for member price and you cloud save 10% or more on thousands of hotel</p>
+        <div className='input-wrapper'>
+          <input type='text' placeholder='Your email address'/>
+          <button>Signup</button>
+        </div>
+      </div>
+    </section>
+  )
+}
